Remove redundant tests for SpriteKit
These exact tests already exist in
validation-test/stdlib/SpriteKit.swift.
